composer: Add 100ms delay upon releasing of large comp hint
When HWC detects that the large composition cycle ends, it releases
the hint only after a certain number of frames depending on the
active refresh rate. This caused HWC to hold the CPU/GPU resources
longer than it should.

To avoid such scenarios, this change releases the hint 100ms after
the large composition cycle regardless of the refresh rate.

Change-Id: Iacdaa600c2c45af24e97baf5a73783646fa96aef